# [wasm] Implement INITBLK and CPBLK in the Jiterpreter