This addresses some benchmark regressions and should provide a speed-up over baseline:

|               Method | Before   | After    |
|--------------------- |---------:|---------:|
| InitBlockAllZeros128 | 22.19 ns | 6.325 ns |
|  InitBlockAllOnes128 | 21.62 ns | 6.236 ns |
|         CopyBlock128 | 23.70 ns | 9.593 ns |